I will not use == for strings.
I will not use == for strings. I will not use == for strings. I will not use == for strings. I will not use == for strings. I will not use == for strings. I will not use == for strings. I will not use == for strings. I will not use == for strings. I will not use == for strings. I will not use == for strings. I will not use == for strings. I will not use == for strings. I will not use == for strings.