Start removing completeAfter()
This function causes games in other channels to lag for no apparent reason, so the bot will need rewriting to not use it pretty much anywhere. The good news is that this bugfix works - the bad news is that this only clears out the minigame lag (as this is where it's at its worst), and the rest will need fixing at another time.