Add forgotten betting check
Right now the bet command never actually checks to make sure it's a betting channel. This leads to an exception if someone tries it (oops). While the exception doesn't cause too much trouble, it's still MUCH better to display a message to the player rather than dumping the stack to the console.