Updated Games enum
Added Call Your Shot to the enum, picked two games to be rotated out to keep the total in the pool at 12, and added the SBR as well so that the Minigame Test can pull from it.

For others: The easiest way to avoid SBR-related compilation errors is to copy another minigame's file into its slot.